creatint list

+++ pushswap.h
@@ -6,7 +6,7 @@
 /*   By: riwatana <riwatana@student.42.jp>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/18 21:54:34 by riwatana          #+#    #+#             */
-/*   Updated: 2026/05/19 20:18:26 by riwatana         ###   ########.fr       */
+/*   Updated: 2026/05/19 23:00:20 by kkomurat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,11 @@
 #include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <unistd.h>
 
 typedef struct s_node
 {
-	int				value;
+	long			value;
 	struct s_node	*next;
 	struct s_node	*prev;
